pin version of webpack-dev-server to avoid build issues
Ran into some issues running the UI locally which seem to related
to webpack-dev-server being unintentially upgraded to a 4.x version.

Pinning the version to  "webpack-dev-server": "3.11.3" seems to fix those issue.

Fixing NPM build issues is by no means my forte, so it's possible this change is the
wrong way to go.