Remove suport for onlyShowActivelyRunning URL request parameter
The ability to switch modes via URL parameter was handy, but there was an issue with the routing in staging/prod
that caused it not to work.
After having a play around, it doesn't feel high value to try to fix it, so we'll just link directly to the
list and the user can just click the check box.

Makes sense to simplify the code and remove it now that it's not going to be used