Add loggings and allow 0 as range start for s3 backfills (#438)
Revamp S3 range validation with detailed message and allow 0 as the
range start value.

For S3 Backfills -
The range isn’t actually defaulting to 0 — what’s happening is that it’s
cloning the values from the previous run. For example, [this
run](https://backfila.sqprod.co/backfills/28700) has a range of 0 to
null. So when cloned, the start value of 0 gets copied over. On the
other hand, [this run](https://backfila.sqprod.co/backfills/28806) has
both start and end as null, so those values are cloned instead.

This means that if a previous run had a start value of 0, and the user
selects RESTART, that 0 will be reused — which can will cause errors for
S3 since valid range should not be there. To avoid this, We want to
support a start value of 0 for S3 backfills.

current logging user get - 
![Screenshot 2025-04-17 at 2 21
46 PM](https://github.com/user-attachments/assets/00b768c2-bd6d-47dd-bed6-e98e193536ee)