Enable Soft Deletion Backfills (#443)
Allow soft-delete backfills for cancelled and completed backfills. 

Soft Deletion Handling: Backfila is updated to recognize and process
records flagged as soft-deleted, ensuring they are included or excluded
from backfills based on configuration.​

A 'Delete' button has been added to cancelled or completed backfills in
the UI, providing users with the ability to soft delete backfills.

We filter the deleted backfills from the UI.



https://github.com/user-attachments/assets/2f755a1f-b0c7-4b95-a515-8b6d000dc320

Modified to nullable timestamp
![Screenshot 2025-04-29 at 1 01
39 PM](https://github.com/user-attachments/assets/7cbd8040-9633-4fa3-baf4-caa9ae227e78)