Add last registered at column to Services (#458)
We're adding a new last_registered_at column to the services table. This
is similar to the updated_at column in registered_backfills, but instead
of tracking only config changes, it will also update whenever the
service is re-registered.
This column will be used in our deprecation reminder. 

When reregistered
![Screenshot 2025-05-14 at 3 27
40 PM](https://github.com/user-attachments/assets/bff3e475-2640-434f-af36-5e16b95af61e)

When registered
![Screenshot 2025-05-14 at 3 27
17 PM](https://github.com/user-attachments/assets/15828ba5-a837-4807-b121-474df8a85fa4)