Add JvmOverloads to BackfillRegistration for binary compatibility (#467)
The addition of the `unit` argument recently is leading to
`NoSuchMethodError`s where code is compiled against an old versio of
backfila without this argument, but a new version is present at runtime
with the argument. This might fix it by adding overloaded methods
(although, I'm not entirely sure whether the new code will be compatible
with the code compiled against the method without the JvmOverloads
annotation), but it'll at least prevent similar errors in future.