[client-sqldelight-gradle-plugin] Update the generated methods and queries to be prefixed by the record source config name (#506)
## Description

This PR comes about from me running into issues with how backfila +
sqldelight interact with each other and the generated code not doing
_quite_ what I'd want it to do.

The TL;DR of my issue can be boiled down to:
* I have a table `users` with a column `phone_number`.
* The type of `phone_number` is `String?`
* I am trying to add a new record source that uses `whereClause =
"phone_number IS NOT NULL"` in the config
* During code generation, sqldelight sees that `phone_number` no longer
maps to the `users` table as the type is now `String` (because
`whereClause = "phone_number IS NOT NULL"` enforces that `phone_number`
couldn't possibly be null)
* Because of this sqldelight creates a new record type matching the
query name. In the case of the generated code from the gradle plugin,
this type would be `GetBatch`.
* While I _could_ use `GetBatch` as the type, this is not very intuitive
or "useful" (you wouldn't call and apple an orange).

## Solution

This PR updates the code generation to prefix the `getBatch` method and
query with the record source config name.

So given the following
```kotlin
addRecordSource(
    name = "usersWithPhoneNumbers",
    ...
    whereClause = "phone_number IS NOT NULL"
)
```
The generated query would be `usersWithPhoneNumbersGetBatch` and the
record type would be `UsersWithPhoneNumbersGetBatch`.

Not only does this change cover the `where x is not null` case, it also
covers cases where you specify specific columns in the `recordColumns`
field.