Make parameters classes easier to work with
The key change is changing the type of
`BackfillRegistration.parametersClass` from `KClass<Any>` to `KClass<out
Any>`. This is a clearer expression of the intent to allow backfill
parameters to be of any type.

In the previous implementation, the type was `KClass<Any>` which
actually means the class of the `Any` type. This required some
gymnastics for the standard clients to be able to register themselves
because they had to cast their parameters type to `Any` (usually by
casting the backfill class to something like `SomeBackfill<Any, Any>` so
that a parameters type of `Any` gets extracted).

With this change, a lot of these type gymnastics can be simplified. This
also makes it much more straightforward to implement a custom client
that uses a custom parameters type.